Re-arrange MonadFilesys
Want to separate a filesystem implementation (a bag of operations) from
the capability to get a filesystem. Because there are a lot of
filesystem operations it makes sense not to write a forwarding typeclass
instance every time some ReaderT Env includes a filesystem, but this
does result in some upfront boilerplate.